Implement / standardize validation in the view

+++ starter-master/src/com/uca/gui/AwardGUI.java
@@ -14,25 +14,38 @@ import java.sql.Date;
 import java.util.HashMap;
 import java.util.Map;
 
+import static com.uca.util.IDUtil.isValidId;
+import static com.uca.util.StringUtil.isValidString;
+
 public class AwardGUI extends _BasicGUI
 {
+    private static boolean isByStudent;
+
     public static String create(String motive, String teacherUserName, long studentId, long stickerId)
             throws IOException, TemplateException
     {
-        AwardEntity award = new AwardEntity();
-        award.setAttributionDate(new Date(new java.util.Date().getTime()));
-        award.setMotive(motive);
-        award.setTeacher(TeacherCore.readByUserName(teacherUserName));
-        award.setStudent(StudentCore.readById(studentId));
-        award.setSticker(StickerCore.readById(stickerId));
-        infoMsg = AwardCore.create(award) != null ? InfoMsg.AJOUT_SUCCES : InfoMsg.AJOUT_ECHEC;
-
+        if (!isValidId(stickerId) || !isValidId(studentId))
+        {
+            infoMsg = InfoMsg.ID_INVALIDE;
+        }
+        else
+        {
+            if (!isValidString(motive) || !isValidString(teacherUserName))
+            {
+                infoMsg = InfoMsg.CHAMPS_VIDES_INTERDITS;
+            }
+            else
+            {
+                AwardEntity award = new AwardEntity();
+                award.setAttributionDate(new Date(new java.util.Date().getTime()));
+                award.setMotive(motive);
+                award.setTeacher(TeacherCore.readByUserName(teacherUserName));
+                award.setStudent(StudentCore.readById(studentId));
+                award.setSticker(StickerCore.readById(stickerId));
+                infoMsg = AwardCore.create(award) != null ? InfoMsg.AJOUT_SUCCES : InfoMsg.AJOUT_ECHEC;
+            }
+        }
         return readAll(true);
-    }
-
-    public static String readAll(boolean isAuthorized) throws IOException, TemplateException
-    {
-        return readMany(isAuthorized, -1);
     }
 
     /**
@@ -49,16 +62,24 @@ public class AwardGUI extends _BasicGUI
     {
         Map<String, Object> input    = new HashMap<>();
         Template            template = _FreeMarkerInitializer.getContext().getTemplate("awards/awards.ftl");
+        input.put("isByStudent", isByStudent);
 
-        if (studentId <= 0)
+        if (!isByStudent)
         {
             input.put("awards", AwardCore.readAll());
-            input.put("isByStudent", false);
         }
         else
         {
-            input.put("isByStudent", true);
-            input.put("awards", AwardCore.readByStudentId(studentId));
+            if (!isValidId(studentId))
+            {
+                infoMsg = InfoMsg.ID_INVALIDE;
+                isByStudent = false;
+                return readMany(isAuthorized, studentId);
+            }
+            else
+            {
+                input.put("awards", AwardCore.readByStudentId(studentId));
+            }
         }
         if (isAuthorized)
         {
@@ -72,7 +93,14 @@ public class AwardGUI extends _BasicGUI
     public static String readByStudentId(boolean isAuthorized, long studentId)
             throws IOException, TemplateException
     {
+        isByStudent = true;
         return readMany(isAuthorized, studentId);
+    }
+
+    public static String readAll(boolean isAuthorized) throws IOException, TemplateException
+    {
+        isByStudent = false;
+        return readMany(isAuthorized, 0);
     }
 
     public static String readById(boolean isAuthorized, long id) throws IOException, TemplateException
@@ -80,15 +108,30 @@ public class AwardGUI extends _BasicGUI
         Map<String, Object> input    = new HashMap<>();
         Template            template = _FreeMarkerInitializer.getContext().getTemplate("awards/award.ftl");
 
-        input.put("award", AwardCore.readById(id));
-        input.put("isAuthorized", isAuthorized);
+        if (!isValidId(id))
+        {
+            infoMsg = InfoMsg.ID_INVALIDE;
+        }
+        else
+        {
+            input.put("award", AwardCore.readById(id));
+            input.put("isAuthorized", isAuthorized);
+        }
         return render(template, input, new StringWriter());
     }
 
     public static String deleteById(long id) throws IOException, TemplateException
     {
-        AwardCore.deleteById(id);
+        if (!isValidId(id))
+        {
+            infoMsg = InfoMsg.ID_INVALIDE;
+        }
+        else
+        {
+            AwardCore.deleteById(id);
+        }
         // we assume that the user was only able to access this function because it was authorized
-        return readMany(true, -1);
+        isByStudent = false;
+        return readMany(true, 0);
     }
 }
