+++ starter-master/src/com/uca/gui/StudentGUI.java
@@ -10,15 +10,25 @@ import java.io.StringWriter;
 import java.util.HashMap;
 import java.util.Map;
 
+import static com.uca.util.IDUtil.isValidId;
+import static com.uca.util.StringUtil.isValidString;
+
 public class StudentGUI extends _BasicGUI
 {
     public static String create(String firstName, String lastName)
             throws IOException, TemplateException
     {
-        StudentEntity student = new StudentEntity();
-        student.setFirstName(firstName);
-        student.setLastName(lastName);
-        infoMsg = StudentCore.create(student) != null ? InfoMsg.AJOUT_SUCCES : InfoMsg.AJOUT_ECHEC;
+        if (!isValidString(firstName) || !isValidString(lastName))
+        {
+            infoMsg = InfoMsg.CHAMPS_VIDES_INTERDITS;
+        }
+        else
+        {
+            StudentEntity student = new StudentEntity();
+            student.setFirstName(firstName);
+            student.setLastName(lastName);
+            infoMsg = StudentCore.create(student) != null ? InfoMsg.AJOUT_SUCCES : InfoMsg.AJOUT_ECHEC;
+        }
 
         return readAll();
     }
@@ -36,26 +46,53 @@ public class StudentGUI extends _BasicGUI
     {
         Map<String, Object> input    = new HashMap<>();
         Template            template = _FreeMarkerInitializer.getContext().getTemplate("students/student.ftl");
-
-        input.put("student", StudentCore.readById(id));
+        if (!isValidId(id))
+        {
+            infoMsg = InfoMsg.ID_INVALIDE;
+        }
+        else
+        {
+            input.put("student", StudentCore.readById(id));
+        }
         return render(template, input, new StringWriter());
     }
 
     public static String update(long id, String firstName, String lastName)
             throws IOException, TemplateException
     {
-        StudentEntity student = new StudentEntity();
-        student.setId(id);
-        student.setFirstName(firstName);
-        student.setLastName(lastName);
-        infoMsg = StudentCore.update(student, id) != null ? InfoMsg.MODIFICATION_SUCCES : InfoMsg.MODIFICATION_ECHEC;
-
+        if (!isValidId(id))
+        {
+            infoMsg = InfoMsg.ID_INVALIDE;
+        }
+        else
+        {
+            if (!isValidString(firstName) || !isValidString(lastName))
+            {
+                infoMsg = InfoMsg.CHAMPS_VIDES_INTERDITS;
+            }
+            else
+            {
+                StudentEntity student = new StudentEntity();
+                student.setId(id);
+                student.setFirstName(firstName);
+                student.setLastName(lastName);
+                infoMsg = StudentCore.update(student, id) != null ? InfoMsg.MODIFICATION_SUCCES
+                                                                  : InfoMsg.MODIFICATION_ECHEC;
+            }
+        }
         return readById(id);
     }
 
     public static String deleteById(long id) throws TemplateException, IOException
     {
-        StudentCore.deleteById(id);
+        if (!isValidId(id))
+        {
+            infoMsg = InfoMsg.ID_INVALIDE;
+        }
+        else
+        {
+            StudentCore.deleteById(id);
+        }
         return readAll();
     }
 }
